feat(core): add `graph` mod and `GraphBuilder`
