+++ api-server/llama-core/src/lib.rs
@@ -9,25 +9,25 @@ pub mod chat;
 pub mod completions;
 pub mod embeddings;
 pub mod error;
+pub mod graph;
 pub mod images;
 pub mod models;
 pub mod rag;
 pub mod utils;
 
 pub use error::LlamaCoreError;
+pub use graph::{EngineType, Graph, GraphBuilder};
 
 use chat_prompts::PromptTemplateType;
 use once_cell::sync::OnceCell;
 use serde::{Deserialize, Serialize};
 use std::{
     collections::HashMap,
+    path::Path,
     sync::{Mutex, RwLock},
 };
-use utils::{get_output_buffer, set_tensor_data_u8};
+use utils::get_output_buffer;
 use wasmedge_stable_diffusion::*;
-use wasmedge_wasi_nn::{
-    Error as WasiNnError, Graph as WasiNnGraph, GraphExecutionContext, TensorType,
-};
 
 // key: model_name, value: Graph
 pub(crate) static CHAT_GRAPHS: OnceCell<Mutex<HashMap<String, Graph>>> = OnceCell::new();
@@ -322,163 +322,6 @@ impl AudioMetadataBuilder {
 
     pub fn build(self) -> Metadata {
         self.metadata
-    }
-}
-
-/// Wrapper of the `wasmedge_wasi_nn::Graph` struct
-#[derive(Debug)]
-pub struct Graph {
-    pub created: std::time::Duration,
-    pub metadata: Metadata,
-    _graph: WasiNnGraph,
-    context: GraphExecutionContext,
-}
-impl Graph {
-    /// Create a new computation graph from the given metadata.
-    pub fn new(metadata: &Metadata) -> Result<Self, LlamaCoreError> {
-        let config = serde_json::to_string(&metadata).map_err(|e| {
-            let err_msg = e.to_string();
-
-            #[cfg(feature = "logging")]
-            error!(target: "stdout", "{}", &err_msg);
-
-            LlamaCoreError::Operation(err_msg)
-        })?;
-
-        // load the model
-        let graph = wasmedge_wasi_nn::GraphBuilder::new(
-            wasmedge_wasi_nn::GraphEncoding::Ggml,
-            wasmedge_wasi_nn::ExecutionTarget::AUTO,
-        )
-        .config(config)
-        .build_from_cache(&metadata.model_alias)
-        .map_err(|e| {
-            let err_msg = e.to_string();
-
-            #[cfg(feature = "logging")]
-            error!(target: "stdout", "{}", &err_msg);
-
-            LlamaCoreError::Operation(err_msg)
-        })?;
-
-        // initialize the execution context
-        let context = graph.init_execution_context().map_err(|e| {
-            let err_msg = e.to_string();
-
-            #[cfg(feature = "logging")]
-            error!(target: "stdout", "{}", &err_msg);
-
-            LlamaCoreError::Operation(err_msg)
-        })?;
-
-        let created = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map_err(|e| {
-                let err_msg = e.to_string();
-
-                #[cfg(feature = "logging")]
-                error!(target: "stdout", "{}", &err_msg);
-
-                LlamaCoreError::Operation(err_msg)
-            })?;
-
-        Ok(Self {
-            created,
-            metadata: metadata.clone(),
-            _graph: graph,
-            context,
-        })
-    }
-
-    /// Get the name of the model
-    pub fn name(&self) -> &str {
-        &self.metadata.model_name
-    }
-
-    /// Get the alias of the model
-    pub fn alias(&self) -> &str {
-        &self.metadata.model_alias
-    }
-
-    /// Get the prompt template type
-    pub fn prompt_template(&self) -> PromptTemplateType {
-        self.metadata.prompt_template
-    }
-
-    /// Update metadata
-    pub fn update_metadata(&mut self) -> Result<(), LlamaCoreError> {
-        #[cfg(feature = "logging")]
-        info!(target: "stdout", "Update metadata for the model named {}", self.name());
-
-        // update metadata
-        let config = match serde_json::to_string(&self.metadata) {
-            Ok(config) => config,
-            Err(e) => {
-                let err_msg = format!("Failed to update metadta. Reason: Fail to serialize metadata to a JSON string. {}", e);
-
-                #[cfg(feature = "logging")]
-                error!(target: "stdout", "{}", &err_msg);
-
-                return Err(LlamaCoreError::Operation(err_msg));
-            }
-        };
-
-        let res = set_tensor_data_u8(self, 1, config.as_bytes());
-
-        #[cfg(feature = "logging")]
-        info!(target: "stdout", "Metadata updated successfully.");
-
-        res
-    }
-
-    /// Set input uses the data, not only [u8](https://doc.rust-lang.org/nightly/std/primitive.u8.html), but also [f32](https://doc.rust-lang.org/nightly/std/primitive.f32.html), [i32](https://doc.rust-lang.org/nightly/std/primitive.i32.html), etc.
-    pub fn set_input<T: Sized>(
-        &mut self,
-        index: usize,
-        tensor_type: TensorType,
-        dimensions: &[usize],
-        data: impl AsRef<[T]>,
-    ) -> Result<(), WasiNnError> {
-        self.context.set_input(index, tensor_type, dimensions, data)
-    }
-
-    /// Compute the inference on the given inputs.
-    pub fn compute(&mut self) -> Result<(), WasiNnError> {
-        self.context.compute()
-    }
-
-    /// Compute the inference on the given inputs.
-    ///
-    /// Note that this method is used for the stream mode. It generates one token at a time.
-    pub fn compute_single(&mut self) -> Result<(), WasiNnError> {
-        self.context.compute_single()
-    }
-
-    /// Copy output tensor to out_buffer, return the output’s **size in bytes**.
-    pub fn get_output<T: Sized>(
-        &self,
-        index: usize,
-        out_buffer: &mut [T],
-    ) -> Result<usize, WasiNnError> {
-        self.context.get_output(index, out_buffer)
-    }
-
-    /// Copy output tensor to out_buffer, return the output’s **size in bytes**.
-    ///
-    /// Note that this method is used for the stream mode. It returns one token at a time.
-    pub fn get_output_single<T: Sized>(
-        &self,
-        index: usize,
-        out_buffer: &mut [T],
-    ) -> Result<usize, WasiNnError> {
-        self.context.get_output_single(index, out_buffer)
-    }
-
-    /// Clear the computation context.
-    ///
-    /// Note that this method is used for the stream mode. It clears the context after the stream mode is finished.
-    pub fn finish_single(&mut self) -> Result<(), WasiNnError> {
-        self.context.fini_single()
     }
 }
 
@@ -898,12 +741,29 @@ pub fn init_stable_diffusion_context(gguf: impl AsRef<str>) -> Result<(), LlamaC
 }
 
 /// Initialize the whisper context
-pub fn init_audio_context(metadata: &Metadata) -> Result<(), LlamaCoreError> {
+pub fn init_whisper_context(
+    metadata: &Metadata,
+    model_file: impl AsRef<Path>,
+) -> Result<(), LlamaCoreError> {
     #[cfg(feature = "logging")]
     info!(target: "stdout", "Initializing the audio context");
 
+    let model_bin = std::fs::read(model_file.as_ref()).map_err(|e| {
+        let err_msg = format!("Failed to read the model. Reason: {}", e);
+
+        #[cfg(feature = "logging")]
+        error!(target: "stdout", "{}", err_msg);
+
+        LlamaCoreError::InitContext(err_msg)
+    })?;
+    info!(target: "stdout", "Read model, size in bytes: {}", model_bin.len());
+
     // create and initialize the audio context
-    let graph = Graph::new(metadata)?;
+    let graph = GraphBuilder::new(EngineType::Whisper)?
+        .with_config(&metadata)?
+        .use_cpu()
+        .build_from_files(&[model_file.as_ref()])?;
+
     AUDIO_GRAPH.set(Mutex::new(graph)).map_err(|_| {
             let err_msg = "Failed to initialize the audio context. Reason: The `AUDIO_GRAPH` has already been initialized";
